user verification on delete + user extracting middleware

+++ osa4/bloglist/controllers/blogRouters.js
@@ -2,19 +2,17 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, id: 1, name: 1 })
     response.json(blogs)
   })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor ,async (request, response) => {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-      return response.status(401).json( {error: 'token invalid'} )
-    }
-    const user = await User.findById(decodedToken.id)
+    
+    const user = request.user
 
     const blog = new Blog(
       {
@@ -33,9 +31,19 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
   })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  } else {
+    return response.status(403).json({ error: 'permission denied' })
+  }
+  
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
